requirements.txt
Pyqt5
Pillow
oead
Pymsyt
bs4
pyqt5-tools
requests
pywin32